+++ packages/libs/common/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./logging";
-export * from "./yaml-parsing";
+export * from "./literate-yaml";
 export * from "./exceptions";
 export * from "./merging";
