+++ packages/libs/common/src/literate-yaml/error-codes.ts
@@ -1,0 +1,7 @@
+/**
+ * Mapping of literate yaml error codes.
+ */
+export const LiterateYamlErrorCodes = {
+  jsonParsingError: "literate/json-parsing-error",
+  yamlParsingError: "literate/yaml-parsing-error",
+};
