+++ packages/libs/common/src/literate-yaml/guard.ts
